Rename next-api wp to wordpress, updates refs
