+++ pages/blog/[[...slug]].js
@@ -7,7 +7,7 @@ import Blocks from '@/components/molecules/Blocks'
 import Card from '@/components/molecules/Card'
 import Comments from '@/components/molecules/Comments'
 import getPagePropTypes from '@/functions/getPagePropTypes'
-import getArchivePosts from '@/lib/next-api/wp/archive/getArchivePosts'
+import getArchivePosts from '@/lib/next-api/wordpress/archive/getArchivePosts'
 import getPostTypeStaticPaths from '@/lib/wordpress/_global/getPostTypeStaticPaths'
 import getPostTypeStaticProps from '@/lib/wordpress/_global/getPostTypeStaticProps'
